requirements.txt
# Core ML Libraries
# For CPU-only (smaller download, ~100MB vs 2GB+):
# Install with: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
# Or uncomment below for GPU version (requires CUDA and more disk space):
# torch>=2.0.0            # PyTorch for deep learning (AST model) - GPU version
# torchaudio>=2.0.0       # PyTorch audio processing utilities - GPU version
# tensorflow>=2.13.0    # Not needed for current implementation
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.0.0           # For model saving/loading

# Transformers & Vision Models
transformers>=4.30.0    # Hugging Face transformers (for ViT models)
pillow>=10.0.0          # Image processing for spectrograms

# Audio Processing
# Required for AST model training with raw audio processing
librosa>=0.10.0         # Audio analysis and feature extraction
soundfile>=0.12.0       # Audio file I/O
audioread>=3.0.0        # Audio file decoding

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter Notebooks
# Only needed if you're working with notebooks
# jupyter>=1.0.0
# ipykernel>=6.25.0
# notebook>=7.0.0

# Model Training & Monitoring
# tensorboard>=2.13.0    # Only needed for neural network training
# wandb>=0.15.0          # Only needed for neural network training
tqdm>=4.65.0             # Useful for progress bars

# Data Handling
h5py>=3.9.0
hdf5storage>=0.1.18

# Utilities
scipy>=1.11.0
joblib>=1.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Optional: For advanced audio analysis - not needed for current implementation
# madmom>=0.16.1
# essentia>=2.1b6.dev858

# Development Tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Presentation Generation
python-pptx>=0.6.21
mistune>=2.0.0
